Deployment

+++ app.py
@@ -1,4 +1,4 @@
-from flask import Flask, render_template, request, redirect, url_for, session, flash
+from flask import Flask, render_template, request, redirect, url_for, session, flash, abort
 from flask_bootstrap import Bootstrap5
 from datetime import datetime, timezone
 from pymongo import MongoClient
@@ -7,12 +7,14 @@ import gridfs
 import markdown
 from flask_login import UserMixin, LoginManager
 from flask_login import login_required, current_user, login_user, logout_user
-from forms import BlogPostForm, LoginForm, RegisterForm, ProfileForm
+from forms import BlogPostForm, LoginForm, RegisterForm, ProfileForm, UserForm
 from werkzeug.security import generate_password_hash, check_password_hash
 from itsdangerous import URLSafeTimedSerializer
 from flask_mail import Mail, Message
 from dotenv import load_dotenv
 import os
+from flask_principal import Principal, Permission, RoleNeed, Identity, identity_changed, identity_loaded, UserNeed, Need
+
 
 app = Flask(__name__)
 app.secret_key = os.getenv('SECRET_KEY')  # Ovo postavljamo kao tajni ključ za sigurnost
@@ -24,7 +26,8 @@ app.config['MAIL_PASSWORD'] = os.getenv('MAIL_PASSWORD')
 app.config['MAIL_DEFAULT_SENDER'] = os.getenv('MAIL_DEFAULT_SENDER')
 
 bootstrap = Bootstrap5(app)
-client = MongoClient('mongodb://localhost:27017/')
+# client = MongoClient('mongodb://localhost:27017/')
+client = MongoClient(os.getenv('MONGODB_CONNECTION_STRING'))
 db = client['pzw_blog_database']
 posts_collection = db['posts']
 users_collection = db['users']
@@ -35,16 +38,23 @@ login_manager = LoginManager()
 login_manager.init_app(app)
 login_manager.login_view = 'login'
 
+principal = Principal(app)
+admin_permission = Permission(RoleNeed('admin'))
+author_permission = Permission(RoleNeed('author'))
+
 @login_manager.user_loader
 def load_user(email):
     user_data = users_collection.find_one({"email": email})
     if user_data:
-        return User(user_data['email'])
+        return User(user_data['email'], user_data.get('is_admin'), user_data.get('theme'))
+
     return None
 
 class User(UserMixin):
-    def __init__(self, email):
+    def __init__(self, email, admin=False, theme=''):
         self.id = email
+        self.admin = admin is True
+        self.theme = theme
 
     @classmethod
     def get(self_class, id):
@@ -52,6 +62,10 @@ class User(UserMixin):
             return self_class(id)
         except UserNotFoundError:
             return None
+        
+    @property
+    def is_admin(self):
+        return self.admin
 
 class UserNotFoundError(Exception):
     pass
@@ -93,11 +107,15 @@ def post_view(post_id):
         flash("Članak nije pronađen!", "danger")
         return redirect(url_for('index'))
 
-    return render_template('blog_view.html', post=post)
+    return render_template('blog_view.html', post=post, edit_post_permission=edit_post_permission)
 
 @app.route('/blog/edit/<post_id>', methods=["get", "post"])
 @login_required
 def post_edit(post_id):
+    permission = edit_post_permission(post_id)
+    if not permission.can():
+        abort(403, "Nemate dozvolu za uređivanje ovog članka.")
+
     form = BlogPostForm()
     post = posts_collection.find_one({"_id": ObjectId(post_id)})
 
@@ -136,6 +154,10 @@ def post_edit(post_id):
 @app.route('/blog/delete/<post_id>', methods=['POST'])
 @login_required
 def delete_post(post_id):
+    permission = edit_post_permission(post_id)
+    if not permission.can():
+        abort(403, "Nemate dozvolu za brisanje ovog posta.")
+
     posts_collection.delete_one({"_id": ObjectId(post_id)})
     flash('Članak je uspješno obrisan.', 'success')
     return redirect(url_for('index'))
@@ -175,6 +197,7 @@ def login():
                 return redirect(url_for('login'))
             user = User(user_data['email'])
             login_user(user, form.remember_me.data)
+            identity_changed.send(app, identity=Identity(user.id))
             next = request.args.get('next')
             if next is None or not next.startswith('/'):
                 next = url_for('index')
@@ -251,47 +274,55 @@ def confirm_email(token):
     
     return redirect(url_for('login'))
 
-@app.route('/profile', methods=['GET', 'POST'])
-@login_required
-def profile():
-    form = ProfileForm(obj=current_user)  # Pre-fill form with current user's data
-    
-    user_data = users_collection.find_one({"email": current_user.get_id()})
-    if request.method == 'GET':
-        form.first_name.data = user_data["first_name"]
-        form.last_name.data = user_data["last_name"]
-        form.bio.data = user_data["bio"]
-    elif form.validate_on_submit():
-        current_user.first_name = form.first_name.data
-        current_user.last_name = form.last_name.data
-        current_user.bio = form.bio.data
 
+def update_user_data(user_data, form):
+    if form.validate_on_submit():
         db.users.update_one(
-        {"email": current_user.get_id()},
+        {"_id": user_data['_id']},
         {"$set": {
-            "first_name": current_user.first_name,
-            "last_name": current_user.last_name,
-            "bio": current_user.bio
+            "first_name": form.first_name.data,
+            "last_name": form.last_name.data,
+            "bio": form.bio.data,
+            "theme": form.theme.data
         }}
         )
         if form.image.data:
             # Pobrišimo postojeću ako postoji
             if hasattr(user_data, 'image_id') and user_data.image_id:
-                fs.delete(current_user.profile_photo_id)
+                fs.delete(user_data.image_id)
             
             image_id = save_image_to_gridfs(request, fs)
             if image_id != None:
                 users_collection.update_one(
-                {"email": current_user.get_id()},
+                {"_id": user_data['_id']},
                 {"$set": {
                     'image_id': image_id,
                 }}
             )
-        
-        flash('Vaši podaci su uspješno spremljeni!', 'success')
-        return redirect(url_for('profile'))
+        flash("Podaci uspješno ažurirani!", "success")
+        return True
+    return False
 
-    return render_template('profile.html', form=form, image_id=user_data["image_id"])
+@app.route('/profile', methods=['GET', 'POST'])
+@login_required
+def profile():
+    user_data = users_collection.find_one({"email": current_user.get_id()})
+    form = ProfileForm(data=user_data)
+    title = "Vaš profil"
+    if update_user_data(user_data, form):
+        return redirect(url_for('profile'))
+    return render_template('profile.html', form=form, image_id=user_data.get("image_id"), title=title)
+
+@app.route('/user/<user_id>', methods=['GET', 'POST'])
+@login_required
+@admin_permission.require(http_exception=403)
+def user_edit(user_id):
+    user_data = users_collection.find_one({"_id": ObjectId(user_id)})
+    form = UserForm(data=user_data)
+    title = "Korisnički profil"
+    if update_user_data(user_data, form):
+        return redirect(url_for('users'))
+    return render_template('profile.html', form=form, image_id=user_data.get("image_id"), title=title)
 
 @app.route("/myposts")
 def my_posts():
@@ -309,3 +340,43 @@ def localize_status(status):
 
 # Registirajmo filter za Jinja-u
 app.jinja_env.filters['localize_status'] = localize_status
+
+# Klasa za definiranje potrebe za uređivanjem članka
+class EditPostNeed2(Need):
+    def __init__(self, post_id):
+        super().__init__('edit_post', post_id)
+
+class EditPostNeed:
+    def __init__(self, post_id):
+        self.method = 'edit_post'
+        self.value = post_id
+
+# Pomoćna metoda za provjeru prava uređivanja
+def edit_post_permission(post_id):
+    return Permission(EditPostNeed(str(ObjectId(post_id))))
+
+@identity_loaded.connect_via(app)
+def on_identity_loaded(sender, identity):
+    if current_user.is_authenticated:
+        identity.user = current_user
+        identity.provides.add(UserNeed(current_user.id))
+        identity.provides.add(RoleNeed('author'))
+        if current_user.is_admin:
+            identity.provides.add(RoleNeed('admin'))
+        # Dodajemo EditPostNeed za svaki članak koji je korisnik kreirao
+        user_posts = posts_collection.find({"author": current_user.get_id()})
+        for post in user_posts:
+            identity.provides.add(EditPostNeed(str(post["_id"])))
+
+
+@app.route('/users', methods=['GET', 'POST'])
+@login_required
+@admin_permission.require(http_exception=403)
+def users():
+    users = users_collection.find().sort("email")
+    return render_template('users.html', users = users)
+
+@app.errorhandler(403)
+def access_denied(e):
+    return render_template('403.html', description=e.description), 403
+
